CHK-137 switching to the Int

+++ app/src/main/java/com/chaikasoft/app/domain/models/OperationSummaryDomain.kt
@@ -6,5 +6,5 @@ data class OperationSummaryDomain(
     val timeIso: String,
     val conductor: ConductorDomain,
     val productLineQuantity: Int, // <-- число РАЗНЫХ товаров в операции
-    val totalPrice: Double        // сумма |impact| * price (для быстрого отображения)
+    val totalPrice: Int        // сумма |impact| * price (для быстрого отображения)
 )
